refactor: nickname 수정확인 users로 이동, 로직수정(#67)
