+++ src/users/dto/check-nickname.dto.ts
