+++ src/users/dto/create-user-profile.dto.ts
@@ -1,4 +1,4 @@
-import { CheckNicknameDto } from './../../auth/dto/check-nickname.dto';
+import { CheckNicknameDto } from './check-nickname.dto';
 import { ApiProperty, PickType } from '@nestjs/swagger';
 import { IsOptional, IsNotEmpty, IsString } from 'class-validator';
 import { Users } from 'src/entities/Users';
